Fix IME crash: lifecycle, ProGuard, and crash logger
- onWindowHidden: use ON_PAUSE instead of ON_STOP to keep
  ComposeView composition alive across hide/show cycles
- onConfigurationChanged: removeView before nulling composeView
- onDestroy: cancel serviceScope after lifecycle cleanup
- KeyboardScreen: Color(Long) direct constructor
- ProGuard: remove overly broad androidx.compose/lifecycle keep
  (BOM bundles its own rules)
- NacreApplication: write crash logs to getExternalFilesDir
- Remove unnecessary storage permissions

+++ ime-core/src/main/kotlin/space/manus/nacre/ime/keyboard/KeyboardScreen.kt
@@ -19,8 +19,8 @@ fun KeyboardScreen(service: NacreInputMethodService) {
     val layerManager = service.layerManager
     val layout = layerManager.currentLayout()
     val theme = ThemeProvider.loadSelectedTheme(service)
-    val bgColor = Color(theme.background.toInt())
-    val accentColor = Color(theme.accent.toInt())
+    val bgColor = Color(theme.background)
+    val accentColor = Color(theme.accent)
 
     // Panel state
     var showClipboard by remember { mutableStateOf(false) }
